requirements_gpu.txt
absl-py==0.9.0
astor==0.8.1
bert-tensorflow==1.0.1
cloudpickle==1.4.1
dataclasses==0.7
decorator==4.4.2
falcon==2.0.0
falcon-cors==1.1.7
falcon-multipart==0.2.0
gast==0.3.3
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
importlib-metadata==1.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.18.5
pandas==1.1.1
protobuf==3.12.2
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-hub==0.8.0
tensorflow-probability==0.7.0
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
